Issue #6861: Symmetry Drawing: Decimal input bugs the symmetry.
Symmetry drawing allows decimal points for positioning the axes. This
has the side effect of of producing an imperfect mirroring when the
position contains a decimal point of .50.

Since it makes little sense to mirror at decimal points, this change
introduces unsigned integers instead of doubles to position the axes.