meson: Add installation tags to installed files
Installation tags[0] are a feature introduced in Meson v0.60.0 allowing
to install only a selected subset of the installable files. Note, this
is not about allowing to compile a subset of the project akin to
invoking a build in a subdirectory when using Autotools.

The tags were assigned according to the parent directory the meson.build
file belongs to. This is quite coarse-grained and ideally would match
the Autotools subdirectory compilation logic. At the moment, Meson does
not offer a way to assign in such a way in a maintainable way and an easy
to remember way. This approach offers a number of tags that are
relatively easy to remember.

A somewhat crude way to see what install targets are available, you can
run the following script in a set-up build directory:

$ cat ./meson-info/intro-install_plan.json | jq .[][].tag | sort | uniq

Ideally, there's no install target without a install tag assigned.
Sadly, currently (Meson v1.8.2) the GNOME module does not assign an
install tag to some VAPI dependencies. I proposed a fix for this [1].

[0] https://mesonbuild.com/Installing.html#installation-tags
[1] https://github.com/mesonbuild/meson/pull/14793