Improve usability of vector layers
This commit is intended as a proof of concept - the solution is hacky. We duplicate the path
when a new vector layer is added and add it as a hidden item. When the property is changed,
the hidden item is updated. This commit also squashes multiple selected paths into one when a
vector layer is created.