gimptextlayout-render.c: Don't apply matrix transformtion
When xres != yres, the transformation has some floating point
rounding errors. this leads to the text being slightly outside the text box.
Since it's just the identity transformation simply not applying it fixes the problem.
see bug #1028.

This is not a fix for the general problem but for the time being,
we don't have any text transformations, so this will do.