app/operations: Add OpenCL implementation for set-alpha
This is the MVC to try out the introduced basic OpenCL infrastructure in
GIMP. The operation itself is very simple and thus was the perfect
candidate for a quick implementation.

As a note for myself and others writing OpenCL implementations in the
future, be mindful of parameters with double type. Support for double is
not present in all drivers and so using floats is something I'd
recommend. For that to work, convert all double parameters to floats
before setting kernel arguments!