app/operations: Add central point for OpenCL implementations
This brings in the same basic infrastructure for writing OpenCL
implementations of operations from GEGL.

GEGL operations expect CL implementations in source form which is then
passed to the driver for compilation. The source is passed as a string
and this infrastructure allows for writing CL code in plain text files
which are automatically "stringified" and made available via header
files.