dialogs: implement "create from clipboard" welcome dialog option
This patch implements a new option in the welcome dialog to create a new
image from the clipbaord. It implements a new GtkMenuButton beside the
Create button that opens a menu with the "From Clipboard" option.
Moreover, it also allows to create a new image from the clipbaord using the
standard keyboard shortcut for `edit-paste-as-new-image` action.