dialogs: hide unicode emojis in welcome dialog on macOS
For some reason (probably related to the wat the font is rendered on macOS),
the unicode emojis in the welcome dialog are not rendered correctly on this OS.
This commit hides them on macOS, so that the dialog looks the same as on
other platforms but without the emojis. This is a temporary fix until we
find a better solution or untile we'll redesign this dialog.