app:operations/layer-modes: Clarify names of GEGL vfuncs
Personally, I find names like "parent", "real_parent" rather confusing
when the class tree is not exactly simple. In the case of layer modes it
is GObject -> GeglOperation -> GeglOperationPointCompose3 -> LayerMode.
And 3 of these have a vfunc called 'process'. I find prepending the
class name to the name of a vfunc implementation to be much more helpful
in understanding which function is which.