Addition Of New Fields And Tabs In Editor Window
This commit was done to test the ability to add new fields and tabs to the Metadata Editor window. This was shown via the adding of a field called "Hellfire" to the Description tab which handles the Xmp.xmpRights.WebStatement tag, and a new tab called New which holds a single field called "The Lone One" which is also linked to the Xmp.xmpRights.WebStatement tag.