Added More Fields To Editor Window
This commit entails the addition of a few more fields on the metadata editor window as the client in the issue board wanted the sub-tags for the license tag to be able to be modified by a user. Also in this commit was the addition of two tags whose absence was unnoticed and changing the capitalization of the cc(Creative Commons) metadata tags. There is one error that was found when GIMP was built which is that cc metadata tags throw an error when they are trying to be added with the error saying that there is no namespace info for the cc prefix.